[DEL] crm_claim_rma_type module is deleted and passed to crm_claim_rma
