+++ crm_claim_rma/crm_claim_rma_data.xml
@@ -88,5 +88,23 @@
             <field name="section_id" ref="section_after_sales_service"/>
             <field name="object_id" search="[('model','=','crm.claim')]" model="ir.model"/>
         </record>
+
+
+    <!-- Claims Type -->
+    <record id="crm_claim_type_supplier" model="crm.claim.type">
+        <field name="name">Supplier</field>
+        <field name="active">1</field>
+    </record>
+
+    <record id="crm_claim_type_customer" model="crm.claim.type">
+        <field name="name">Customer</field>
+        <field name="active">1</field>
+    </record>
+
+    <record id="crm_claim_type_other" model="crm.claim.type">
+        <field name="name">Other</field>
+        <field name="active">1</field>
+    </record>
+
     </data>
 </openerp>
