+++ crm_claim_rma/crm_claim_rma_view.xml
@@ -279,7 +279,7 @@
             <field name="arch" type="xml">
                 <xpath expr="//sheet/group[2]" position="replace">
                     <group colspan="4" col="4"  groups="base.group_user">
-                        <field name="claim_type"/>
+                        <field name="claim_type" widget="selection" required="1"/>
                         <field name="warehouse_id" />
                         <field name="user_id" context="{'default_groups_ref': ['base.group_user', 'base.group_partner_manager', 'base.group_sale_salesman_all_leads']}"/>
                         <field name="priority" widget="priority"/>
@@ -375,5 +375,57 @@
             parent="base.menu_aftersale" action="act_crm_case_claim_lines" sequence="2"/>
         <menuitem name="Returned line substates" id="menu_crm_case_claims_claim_line_substates"
             parent="crm_claim.menu_config_claim" action="act_crm_claim_substates" sequence="2"/>
+
+
+        <record id="crm_claim_type_tree" model="ir.ui.view">
+            <field name="name">crm.claim.type.tree</field>
+            <field name="model">crm.claim.type</field>
+            <field name="arch" type="xml">
+                <tree string="Claim types">
+                    <field name="name"/>
+                    <field name="active"/>
+                </tree>
+            </field>
+        </record>
+
+        <record id="crm_claim_type_form" model="ir.ui.view">
+            <field name="name">crm.claim.type.form</field>
+            <field name="model">crm.claim.type</field>
+            <field name="arch" type="xml">
+                <form string="Claim type">
+                    <group col="4">
+                        <field name="name"/>
+                        <field name="active"/>
+                        <field name="description"/>
+                    </group>
+                </form>
+            </field>
+        </record>
+
+        <record id="crm_claim_type_act" model="ir.actions.act_window">
+            <field name="name">Claim Types</field>
+            <field name="res_model">crm.claim.type</field>
+            <field name="view_type">form</field>
+            <field name="view_mode">tree,form</field>
+        </record>
+
+        <menuitem action="crm_claim_type_act" name="Claim Types"
+            id="menu_crm_claim_type_act" parent="crm_claim.menu_config_claim" groups="base.group_no_one"/>
+
+    <record id="crm_stage_claim_type" model="ir.ui.view">
+        <field name="name">crm.claim.stage.type</field>
+        <field name="model">crm.claim.stage</field>
+        <field name="inherit_id" ref="crm_claim.crm_claim_stage_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='sequence']" position="after">
+                <field name="claim_type" widget="selection"/>
+                <!--
+                <field name="claim_default"/>
+                -->
+            </xpath>
+        </field>
+    </record>
+
      </data>
 </openerp>
+
