--- crm_claim_rma_type/README.md
@@ -1,14 +1,0 @@
-CRM RMA Priority Extended
-=========================
-
-To calculate the priority of claim line depends of limit date 
-of warranty
-
-Features
---------
-
-- New field priority in claim line
-- Calculate priority of claim line depending of limit date of warranty
-- Grouping by priority in claim line
-
-
