--- crm_claim_rma_type/data/crm_claim_rma_type.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0"?>
-<openerp>
-    <data noupdate="1">
-        <!-- Claims Sequence n° -->
-        <record id="crm_claim_type_supplier" model="crm.claim.type">
-            <field name="name">Supplier</field>
-            <field name="active">1</field>
-        </record>
-
-        <record id="crm_claim_type_customer" model="crm.claim.type">
-            <field name="name">Customer</field>
-            <field name="active">1</field>
-        </record>
-
-        <record id="crm_claim_type_other" model="crm.claim.type">
-            <field name="name">Other</field>
-            <field name="active">1</field>
-        </record>
-
-    </data>
-</openerp>
