--- crm_claim_rma_type/demo/crm_claim_rma_type.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0"?>
-<openerp>
-    <data noupdate="1">
-
-        <!-- Claims -->
-
-        <record id="crm_claim.crm_claim_1" model="crm.claim">
-            <field name="claim_type" search="[('name', '=', 'Customer')]"/>
-        </record>
-
-        <record id="crm_claim.crm_claim_2" model="crm.claim">
-            <field name="claim_type" ref="crm_claim_type_customer"/>
-        </record>
-
-        <record id="crm_claim.crm_claim_3" model="crm.claim">
-            <field name="claim_type" ref="crm_claim_type_customer"/>
-        </record>
-
-        <record id="crm_claim.crm_claim_4" model="crm.claim">
-            <field name="claim_type" ref="crm_claim_type_customer"/>
-        </record>
-
-        <record id="crm_claim.crm_claim_5" model="crm.claim">
-            <field name="claim_type" ref="crm_claim_type_customer"/>
-        </record>
-
-        <record id="crm_claim.crm_claim_6" model="crm.claim">
-            <field name="claim_type" ref="crm_claim_type_customer"/>
-        </record>
-
-        <record id="crm_claim.crm_claim_7" model="crm.claim">
-            <field name="claim_type" ref="crm_claim_type_customer"/>
-        </record>
-
-        <!-- Claim Stage -->
-
-        <record model="crm.claim.stage" id="crm_claim.stage_claim1">
-            <field name="claim_type" ref="crm_claim_type_customer"/>
-        </record>
-
-        <record model="crm.claim.stage" id="crm_claim.stage_claim5">
-            <field name="claim_type" ref="crm_claim_type_customer"/>
-        </record>
-
-        <record model="crm.claim.stage" id="crm_claim.stage_claim2">
-            <field name="claim_type" ref="crm_claim_type_customer"/>
-        </record>
-
-        <record model="crm.claim.stage" id="crm_claim.stage_claim3">
-            <field name="claim_type" ref="crm_claim_type_customer"/>
-        </record>
-
-
-    </data>
-</openerp>
-
