--- crm_claim_rma_type/view/crm_claim_rma_type.xml
@@ -1,66 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <record id="crm_claim_type_tree" model="ir.ui.view">
-            <field name="name">crm.claim.type.tree</field>
-            <field name="model">crm.claim.type</field>
-            <field name="arch" type="xml">
-                <tree string="Claim types">
-                    <field name="name"/>
-                    <field name="active"/>
-                </tree>
-            </field>
-        </record>
-
-        <record id="crm_claim_type_form" model="ir.ui.view">
-            <field name="name">crm.claim.type.form</field>
-            <field name="model">crm.claim.type</field>
-            <field name="arch" type="xml">
-                <form string="Claim type">
-                    <group col="4">
-                        <field name="name"/>
-                        <field name="active"/>
-                        <field name="description"/>
-                    </group>
-                </form>
-            </field>
-        </record>
-
-        <record id="crm_claim_type_act" model="ir.actions.act_window">
-            <field name="name">Claim Types</field>
-            <field name="res_model">crm.claim.type</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">tree,form</field>
-        </record>
-
-        <menuitem action="crm_claim_type_act" name="Claim Types"
-            id="menu_crm_claim_type_act" parent="crm_claim.menu_config_claim" groups="base.group_no_one"/>
-    <!--
-    <record id="crm_claim_line_type_many2one" model="ir.ui.view">
-        <field name="name">crm.claim.type.many2one</field>
-        <field name="model">crm.claim</field>
-        <field name="inherit_id" ref="crm_claim_rma.crm_claim_rma_form_view2"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='claim_type']" position="attributes">
-                <attribute name="widget">selection</attribute>
-                <attribute name="required">1</attribute>
-            </xpath>
-        </field>
-    </record>
-    -->
-
-    <record id="crm_stage_claim_type" model="ir.ui.view">
-        <field name="name">crm.claim.stage.type</field>
-        <field name="model">crm.claim.stage</field>
-        <field name="inherit_id" ref="crm_claim.crm_claim_stage_form"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='sequence']" position="after">
-                <field name="claim_type" widget="selection"/>
-                <field name="claim_default"/>
-            </xpath>
-        </field>
-    </record>
-
-    </data>
-</openerp>
