+++ rma/__openerp__.py
@@ -42,7 +42,6 @@
         "crm_rma_claim_make_claim",
         "crm_claim_product_supplier",
         "crm_rma_advance_warranty",
-        "crm_claim_rma_type",
     ],
     "demo": [],
     "data": [],
